Move MetaGroup into Elements module
