+++ src/modules/blocks/event-details/block.js
@@ -22,33 +22,12 @@ import {
 	TermsList,
 	OrganizerForm,
 	DateTime,
+	MetaGroup,
 } from 'elements'
 
 /**
  * Module Code
  */
-
-class MetaGroup extends Component {
-	static defaultProps = { className: '', children: null }
-
-	constructor() {
-		super( ...arguments );
-	}
-
-	render() {
-		const { groupKey, className, children } = this.props;
-
-		return (
-			<div
-				className={ `tribe-editor-meta-group tribe-editor-meta-group-${groupKey} ${className}` }
-				key={ groupKey }
-			>
-				{ children }
-			</div>
-		)
-	}
-}
-
 class EventDetails extends Component {
 	constructor() {
 		super( ...arguments );
