+++ src/modules/elements/index.js
@@ -1,3 +1,4 @@
 export { default as TermsList } from './terms-list'
 export { default as OrganizerForm } from './organizer-form'
 export { default as DateTime } from './datetime'
+export { default as MetaGroup } from './datetime'
